+++ src/hooks.tsx
@@ -1,11 +1,11 @@
-import { useContext, useEffect, useState } from 'react';
+import { useCallback, useContext, useEffect, useState } from 'react';
 import { Environment, EnvironmentContext } from './utils/env';
-import { ProjectConfig, ConfigContext } from './utils/projectConfig';
+import { ProjectConfig, ConfigContext, mergeConfig } from './utils/projectConfig';
 import { PageContent, PageContentContext } from './utils/content';
 import { CustomDomain, CustomDomainContext } from './utils/domain';
 import { SlugProperties, SlugPropertiesContext } from './utils/properties';
 import { DebugMode, DebugModeContext } from './utils/debug';
-import { PreviewModeContext, PreviewMode } from './utils/preview';
+import { PreviewModeContext, PreviewMode, buildPreviewProps } from './utils/preview';
 
 export function usePreviewMode(): PreviewMode {
   return useContext(PreviewModeContext);
@@ -70,3 +70,124 @@ export function hasScrolled(y = 0): boolean {
 
   return hasScrolled;
 }
+
+export type FileSystemFileHandles = { [path: string]: FileSystemFileHandle };
+
+async function iterateDirectory(
+  directory: FileSystemDirectoryHandle,
+  relativePath?: string,
+  other?: FileSystemFileHandles,
+): Promise<FileSystemFileHandles> {
+  let handles: FileSystemFileHandles = {
+    ...other,
+  };
+
+  for await (const entry of directory.values()) {
+    if (entry.kind === 'file' && entry.name.endsWith('.mdx')) {
+      handles[`${relativePath ?? ''}/${entry.name.replace('.mdx', '')}`] = entry;
+    }
+
+    if (entry.kind === 'directory') {
+      handles = {
+        ...handles,
+        ...(await iterateDirectory(entry, `${relativePath ?? ''}/${entry.name}`, handles)),
+      };
+    }
+  }
+
+  return handles;
+}
+
+export function useHashChange(): string {
+  const [hash, setHash] = useState('');
+
+  function onHashChange() {
+    // TODO: probably have to handle deeper index files, and redirects somewhere
+    const newHash =
+      window.location.hash.replace('#', '') === '/'
+        ? '/index'
+        : window.location.hash.replace('#', '');
+    return setHash(newHash);
+  }
+
+  useEffect(() => {
+    window.addEventListener('hashchange', onHashChange);
+    return () => window.removeEventListener('hashchange', onHashChange);
+  }, []);
+
+  return hash;
+}
+
+export function useDirectorySelector() {
+  const [error, setError] = useState<Error | null>(null);
+  const [pending, setPending] = useState(false);
+  const [handles, setHandles] = useState<FileSystemFileHandles | null>();
+  const [config, setConfig] = useState<ProjectConfig | null>();
+
+  const select = useCallback(async () => {
+    setPending(true);
+    try {
+      const handle = await window.showDirectoryPicker();
+      let docs: FileSystemDirectoryHandle;
+      let foundDocsJson = false;
+      for await (const entry of handle.values()) {
+        if (entry.kind === 'file' && entry.name === 'docs.json') {
+          const configFile = await entry.getFile();
+          const configText = JSON.parse(await configFile.text());
+
+          setConfig(mergeConfig(configText));
+          foundDocsJson = true;
+        }
+        if (entry.kind === 'directory' && entry.name === 'docs') {
+          docs = entry;
+        }
+      }
+
+      if (!foundDocsJson) {
+        throw new Error('No docs.json found');
+      }
+
+      if (!docs) {
+        throw new Error('No docs directory found');
+      }
+
+      // TODO set config
+      setHandles(await iterateDirectory(docs));
+    } catch (e) {
+      setError(e);
+    } finally {
+      setPending(false);
+    }
+  }, []);
+
+  return { select, handles, error, pending, config };
+}
+
+export const usePollLocalDocs = (handles, config, ms, setPageProps) => {
+  const hash = useHashChange();
+  useEffect(() => {
+    if (!handles) {
+      return;
+    }
+    const handle = hash !== '' ? handles[hash] : handles['/index'];
+
+    // TODO handle no file (404)
+    if (!handle) {
+      console.log('file not found, 404');
+      return;
+    }
+
+    // TODO update state and show page
+    const interval = setInterval(
+      () =>
+        handle
+          .getFile()
+          .then(file => file && file.text())
+          .then(text => text && buildPreviewProps({ hash, config: JSON.stringify(config), text }))
+          .then(props => props && setPageProps(props)),
+      500,
+    );
+
+    return () => clearInterval(interval);
+  }, [handles, hash]);
+};
